python/requirements.txt
python-socketio[client]
numpy
pyserial
requests
pandas
